InfraSpectre/requirements.txt
black==25.1.0
click==8.2.1
grpcio==1.74.0
grpcio-tools==1.74.0
isort==6.0.1
mypy==1.17.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
protobuf==6.31.1
setuptools==80.9.0
typing_extensions==4.14.1
prometheus-client>=0.20,<1
cryptography>=41,<45
PyYAML>=6.0
pytest>=8.3
requests>=2.32.3
